requirements.txt
# python requirements
pip==19.2.3
numpy==1.16.3
pandas==0.24.2
scipy==1.2.1
SimpleITK==1.2.0
Keras==2.2.4
scikit-learn==0.21.1
pyradiomics==2.1.2

# R requirements
psych==1.8.12
combat==2.0
data.table==1.12.6
